pyproject.toml
[tool.poetry]
name = "data-science-repo-template-v2"
version = "0.1.0"
description = ""
authors = ["tom-burdge <tomburdge@wessexinsights.com>"]
readme = "README.md"
packages = [{include = "data_science_repo_template_v2", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
jupyter = "^1.0.0"
seaborn = "^0.12.2"
plotly = "^5.16.0"
scikit-learn = "^1.3.0"
functime = "^0.9.2"
openai = "^1.3.6"
tiktoken = "^0.5.1"
tenacity = "^8.2.3"
streamlit = "^1.28.2"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"
lightgbm = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.284"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
nbqa = "^1.7.0"
nbstripout = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
